ut: reorg, Context manager
