+++ .gitignore
@@ -86,6 +86,7 @@ celerybeat-schedule
 .venv
 venv/
 ENV/
+venv3.6/
 
 # Spyder project settings
 .spyderproject
