+++ coverage.sh
@@ -1,2 +1,2 @@
-TESTING=1 venv/bin/coverage run --source=spruned -m unittest discover
-venv/bin/coverage html
+TESTING=1 venv3.6/bin/coverage run --source=spruned -m unittest discover
+venv3.6/bin/coverage html
