+++ setup3.6.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-rm -rf venv
-virtualenv -p python3.6 venv
-venv/bin/python setup.py install
-venv/bin/pip install coverage==4.5.1
+rm -rf venv3.6
+virtualenv -p python3.6 venv3.6
+venv3.6/bin/python setup.py install
+venv3.6/bin/pip install coverage==4.5.1
 bash ./coverage.sh
