+++ spruned/daemon/bitcoin_p2p/utils.py
@@ -57,18 +57,3 @@ class AsyncBlockFactory:
 
 def get_block_factory():
     return AsyncBlockFactory()
-
-
-def batcher_factory(self):
-    class FakeBatcher:
-        @staticmethod
-        async def add_peer(data):
-            nonlocal self
-            return self.batcher_factory.add_peer(data)
-
-        @staticmethod
-        async def inv_item_to_future(data):
-            nonlocal self
-            return async_coro(self.batcher_factory.inv_item_to_future(data))
-
-    return FakeBatcher()
