+++ test/test_application/test_zeromq.py
@@ -5,6 +5,7 @@ from unittest.mock import Mock, create_autospec
 import async_timeout
 import binascii
 
+import time
 from zmq.asyncio import Context
 import zmq
 import asyncio
@@ -27,7 +28,6 @@ class TestZeroMQ(TestCase):
         self._setup_reactor()
         self._loop = asyncio.get_event_loop()
         self._data_from_topics = {}
-        self.ctx = Context.instance()
 
     def _setup_reactor(self):
         def _save(topic, data):
@@ -70,21 +70,21 @@ class TestZeroMQ(TestCase):
         return block
 
     async def _subscribe_topic(self, topic, url, max_msgs):
-        async with async_timeout.timeout(30):
-            socket = self.ctx.socket(zmq.SUB)
-            socket.connect(url)
-            socket.subscribe(topic)
-            if not self._data_from_topics.get(topic):
-                self._data_from_topics[topic] = []
-            x = 0
-            while x < max_msgs:
-                x += 1
-                print('waiting')
-                msg = await socket.recv_multipart()
-                print('done')
-                self._data_from_topics[topic].append(msg)
-            socket.close()
-            print('socket closed')
+            with Context() as ctx:
+                socket = ctx.socket(zmq.SUB)
+                socket.connect(url)
+                socket.subscribe(topic)
+                if not self._data_from_topics.get(topic):
+                    self._data_from_topics[topic] = []
+                x = 0
+                while x < max_msgs:
+                    x += 1
+                    print('waiting')
+                    msg = await socket.recv_multipart()
+                    print('done')
+                    self._data_from_topics[topic].append(msg)
+                socket.close()
+                print('socket closed')
 
     def test_zmq(self):
         block = self._get_block_with_tx()
@@ -97,9 +97,7 @@ class TestZeroMQ(TestCase):
             self._subscribe_topic(BitcoindZMQTopics.BLOCKHASH.value, self.context.zmqpubhashblock, 1),
             self._test_zmq(block)
         ]
-        self._loop.run_until_complete(
-            asyncio.gather(*_tasks)
-        )
+        self._loop.run_until_complete(asyncio.gather(*_tasks))
 
     async def _test_zmq(self, block):
         await asyncio.sleep(3)
